requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
opencv-python-headless
django-heroku
django
djangorestframework
numpy
scipy
scikit-image
matplotlib
gunicorn
whitenoise